t: remove confusing "! test_grep" invocations
When using the "test_grep" convenience function, callers are encouraged
to write "test_grep ! ..." instead of "! test_grep ..." in order to
assert that a given string is *not* contained in our search.

While both ultimately have the right behavior, writing "! test_grep"
leads to noisy and confusing test output. For instance, if we are
expecting not to find string "foo" in file "bar" and do so by writing

    ! test_grep foo bar

, then (supposing that "bar"'s contents do not in fact contain the
string "foo") the verbose output will contain:

    error: 'grep foo bar' didn't find a match in:
    baz
    ok 1 - test

Even though not finding a match was the expected behavior, test_grep is
unaware and thus we get a confusing "error: " in our verbose output.

Write these invocations instead as "test_grep ! ...", which instructs
test_grep that the caller is expecting not to find a match and ensures
it behaves accordingly.

Signed-off-by: Taylor Blau <me@ttaylorr.com>