t7704: demonstrate repack including new cruft pack in MIDX
When writing a MIDX via 'repack' with the '--write-midx' command-line
option, the documentation in git-repack(1) says:

    Write a multi-pack index (see linkgit:git-multi-pack-index[1])
    containing the non-redundant packs.

Here "non-redundant" means packs that were either just written or will
otherwise not be deleted as a part of the repack operation. However, the
documentation in git-config(1) for 'repack.midxMustContainCruft' says:

    [...] When false, cruft packs are only included in the MIDX
    when necessary (e.g., because they might be required to form a
    reachability closure with MIDX bitmaps). Defaults to true.

But this does not quite match reality. In a '--write-midx' repack where
we wrote one or more cruft pack(s), we will blindly include all of them
in the MIDX operation in which we wrote one or more cruft pack(s).

Because fixing this bug is less than trivial, only demonstrate its
existence in this patch. The immediately following patch fixes the bug.

Signed-off-by: Taylor Blau <me@ttaylorr.com>